Update
A .pdf with the animations demonstrating the functioning of the MLMBio and MLMVol applications. The applications are a technological product of the doctoral thesis entitled: Machine learning to predict biomass and commercial volume of trees in tropical forests. Applications are available at: https://deivisonsouza.shinyapps.io/MLBiomass/ and https://deivisonsouza.shinyapps.io/MLVolume/.